requirements.txt
h5py=3.7.0
ipykernel=6.19.2
ipython=8.9.0
jupyter=1.0.0
lightgbm=3.2.1
matplotlib=3.6.2
numpy=1.23.5
pandas=1.5.2
python=3.10.9
pytorch=1.13.1
scikit-learn=1.2.0
scipy=1.9.3
seaborn=0.11.2
setuptools=65.6.3
shap=0.41.0
six=1.16.0
xgboost=1.7.3